Refactor: Move and rename all operator_pw files to module_pwdft directory and clean up
