+++ source/source_pw/module_pwdft/hamilt_pw.cpp
@@ -8,8 +8,8 @@
 #include "op_pw_ekin.h"
 #include "op_pw_meta.h"
 #include "op_pw_nl.h"
-#include "operator_pw/onsite_proj_pw.h"
-#include "operator_pw/op_exx_pw.h"
+#include "op_pw_proj.h"
+#include "op_pw_exx.h"
 #include "source_hamilt/module_xc/exx_info.h" // use GlobalC::exx_info
 
 namespace hamilt
