+++ source/source_pw/module_pwdft/op_pw_exx.cpp
@@ -1,4 +1,4 @@
-#include "op_exx_pw.h"
+#include "op_pw_exx.h"
 
 #include "source_base/constants.h"
 #include "source_base/global_variable.h"
