+++ source/source_pw/module_pwdft/op_pw_exx_pot.cpp
@@ -1,4 +1,4 @@
-#include "op_exx_pw.h"
+#include "op_pw_exx.h"
 #include "source_io/module_parameter/parameter.h"
 #include "source_hamilt/module_xc/exx_info.h" // use GlobalC::exx_info
