+++ source/source_pw/module_pwdft/op_pw_vel.h
@@ -1,6 +1,6 @@
 #ifndef VELOCITY_PW_H
 #define VELOCITY_PW_H
-#include "operator_pw/operator_pw.h"
+#include "op_pw.h"
 #include "source_cell/unitcell.h"
 #include "source_pw/module_pwdft/VNL_in_pw.h"
 #include "source_basis/module_pw/pw_basis_k.h"
